requirements.txt
# -*- coding: utf-8 -*-

aac-datasets
aac-metrics
black
hydra-core
ipython
lightning
pre-commit
pyyaml
torch
torchoutil
